Minor setup page update
This change was based on feedback that user's were confused by a link within the UI that led to this section. In the UI it talks about not using a supported runtime, but then they are directed to a page on deployment types. My goal is to get them to click their deployment type and then find what they want on the next page. Feel free to reword to help make that more clear!